- version 1.2 beta changes
This is a major overhaul that allows an external program to modify the driver's
settings. The way this works is that when the driver loads, it uses a default
set of settings. A daemon application can detect when a driver loads and call
the IORegistryEntrySetCFProperties() to modify the driver settings. Clients
also can call IORegistryEntryCreateCFProperties() to see what the current
settings are.

• moved property list keys into a DWXBoxHIDDriverKeys.h
• removed device type enumeration - use string instead
• new pad options:
    - bool InvertXAxis;
    - bool InvertRyAxis;
    - bool InvertRxAxis;
    - bool ClampButtons;
    - bool ClampLeftTrigger;
    - bool ClampRightTrigger;
    - UInt8 LeftTriggerThreshold;
    - UInt8 RightTriggerThreshold;
• new methods to handle prefs setting from user space
    - setProperties( OSObject * properties );
    - setDefaultOptions();
    - setupDevice();